requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.2
PyYAML==6.0.1
jinja2==3.1.2
matplotlib==3.8.2
plotly==5.17.0
pandas==2.1.4
numpy==1.26.2
psutil==5.9.6
click==8.1.7
rich==13.7.0
openpyxl==3.1.2
reportlab==4.0.7
aiofiles==23.2.0
tenacity==8.2.3